Restructure LabelDemux Library
Initial changes to LabelDemux library.  It does not build yet.
