+++ .gitignore
@@ -49,3 +49,4 @@
 *.cmake
 *.user
 CMakeCache.txt
+/LabelDemux/build
