+++ LabelDemux/CMakeLists.txt
@@ -1,61 +1,70 @@
-cmake_minimum_required (VERSION 3.8)
-project(LabelDemux)
+cmake_minimum_required (VERSION 3.21)
+project(LabelDemux
+	VERSION 0.2.0
+	DESCRIPTION "Confidentiality Metadata Label Demuxer"
+	LANGUAGES C CXX)
 
-include(GenerateExportHeader)
+# set the postfix "d" for the resulting .so or .dll files when building the
+# library in debug mode
+set(CMAKE_DEBUG_POSTFIX
+    d
+)
 
-# specify the C++ standard
-set(CMAKE_CXX_STANDARD 14)
-set(CMAKE_CXX_STANDARD_REQUIRED True)
-set(CMAKE_CXX_VISIBILITY_PRESET hidden)
-set(CMAKE_VISIBILITY_INLINES_HIDDEN 1)
+add_library(LabelDemux SHARED)
+	
+# set properties for the target. VERSION set the library version to the project
+# version * SOVERSION set the compatibility  version for the library to the
+# major number of the version
+set_target_properties(
+    LabelDemux
+    PROPERTIES VERSION ${PROJECT_VERSION}
+               SOVERSION ${PROJECT_VERSION_MAJOR}
+)
+			   
+file(GLOB SRC_LIST
+	src/*.cpp)
+	
+target_sources(LabelDemux PRIVATE ${SRC_LIST})
+
+# define the C++ standard needed to compile this library and make it visible to
+# dependers
+target_compile_features(
+    LabelDemux
+    PUBLIC cxx_std_17
+)
 
 if (WIN32)
-    set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} /D_USRDLL /D_WINDLL")
-    set(CMAKE_INSTALL_LIBDIR $ENV{APPDATA})
     set(WS "wsock32 ws2_32")
 endif()
 
-if (CMAKE_VS_PLATFORM_NAME)
-    include_directories(../../../../lib/mp2tp/libmp2t)
-    add_library(lcss::libmp2t STATIC IMPORTED)
-    set_target_properties(lcss::libmp2t PROPERTIES
-                 IMPORTED_LOCATION ../../../../lib/mp2tp/libmp2t/Release/libmp2t.lib
-                 IMPORTED_LOCATION_DEBUG ../../../../lib/mp2tp/libmp2t/Debug/libmp2t.lib
-                 IMPORTED_CONFIGURATIONS "RELEASE;DEBUG")
-else ()
-    include(${CMAKE_INSTALL_LIBDIR}/lib/cmake/libmp2t/libmp2tTargets.cmake)
-endif()
-
-file(GLOB SRC_LIST
-    *.h
-    *.cpp)
-add_library(LabelDemux SHARED ${SRC_LIST})
-target_link_libraries(LabelDemux PRIVATE lcss::libmp2t ${WS})
-
-generate_export_header(LabelDemux
-    EXPORT_FILE_NAME "${CMAKE_CURRENT_SOURCE_DIR}/labeldemux_export.h")
-
-target_include_directories(LabelDemux
-          PUBLIC 
-          "$<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}>"
-          "$<INSTALL_INTERFACE:include>"
-          )
-
-install(TARGETS LabelDemux 
-    EXPORT LabelDemuxTargets
-    LIBRARY DESTINATION lib
-    ARCHIVE DESTINATION lib
-    RUNTIME DESTINATION bin
-    INCLUDES DESTINATION include
+# set the include directories
+target_include_directories(
+    LabelDemux
+    PRIVATE src/LabelDemux
+    PUBLIC include
 )
 
-install(FILES
-    LabelDemux.h
-    ${CMAKE_CURRENT_SOURCE_DIR}/labeldemux_export.h  DESTINATION include)
-
-# generate and install export file
-install(EXPORT LabelDemuxTargets
-        FILE LabelDemuxTargets.cmake
-        NAMESPACE ThetaStream::
-        DESTINATION ${CMAKE_INSTALL_LIBDIR}/lib/cmake/LabelDemux
+# if using limited visibility, set CXX_VISIBILILTY_PRESET to "hidden"
+include(GenerateExportHeader)
+set_property(
+    TARGET LabelDemux
+    PROPERTY CXX_VISIBILITY_PRESET "hidden"
+)
+# Hide inlined functions by default, reducing the size of the library
+set_property(
+    TARGET LabelDemux
+    PROPERTY VISIBILITY_INLINES_HIDDEN TRUE
+)
+# this command generates a header file in the CMAKE_CURRENT_BINARY_DIR which
+# sets the visibility attributes according to the compiler settings
+generate_export_header(
+    LabelDemux
+    EXPORT_FILE_NAME
+    export/LabelDemux/export_labeldemux.hpp
+)
+# Add CMAKE_CURRENT_BINARY_DIR to the include path so the generated header can
+# be found
+target_include_directories(
+    LabelDemux
+    PUBLIC "${CMAKE_CURRENT_BINARY_DIR}/export"
 )
