+++ LabelDemux/src/pch.cpp
@@ -1,5 +1,5 @@
 // pch.cpp: source file corresponding to the pre-compiled header
 
-#include "pch.h"
+#include <LabelDemux/pch.h>
 
 // When you are using pre-compiled headers, this source file is necessary for compilation to succeed.
